# docs/swagger.yaml
definitions:
  handlers.BookRequest:
    properties:
      author:
        type: string
      genre:
        type: string
      name:
        type: string
      year:
        type: string
    type: object
  handlers.BookResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.BookResponse'
      message:
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      errorCode:
        type: string
      message:
        type: string
    type: object
  schemas.BookResponse:
    properties:
      author:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      genre:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
      year:
        type: string
    type: object
info:
  contact: {}
paths:
  /book:
    delete:
      consumes:
      - application/json
      description: Delete a book entry
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.BookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete book
      tags:
      - Books
    get:
      consumes:
      - application/json
      description: Show a specific book entry
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.BookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Show book
      tags:
      - Books
    post:
      consumes:
      - application/json
      description: Create a new book entry
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.BookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create book
      tags:
      - Books
  /books:
    get:
      consumes:
      - application/json
      description: List all books
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.BookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List book
      tags:
      - Books
  /random-book:
    get:
      consumes:
      - application/json
      description: Show random entry
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.BookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Random book
      tags:
      - Books
swagger: "2.0"